+++ lib/domain/index.ts
@@ -1,3 +1,4 @@
 export * from "./colorimetry";
 export * from "./qc";
 export * from "./reference-products";
+export * from "./product-mapping";